Move card-wrapper to common quiz module
