+++ app/(language)/language/quiz/quiz-card.tsx
